fix(sdk): restore wrapSigningError as exported function with tests
The linter previously deleted wrapSigningError from signing.ts because
it wasn't re-exported from the barrel. Restore the function, re-export
it from errors/index.ts, restore tests, and use it in
credentials-manager-base.ts instead of the inlined copy.
